expanded object

+++ src/main/java/io/componenttesting/teammorale/dao/TeamsDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface TeamsDao extends JpaRepository<TeamsEntity, Integer> {
+public interface TeamsDao extends JpaRepository<TeamsEntity, Long> {
 
     Optional<TeamsEntity> findByName(String name);
 }
